Fix layering: Move IEditorConstructionOptions to /browser/
